% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_parameters.R
\name{gentic_parameters}
\alias{gentic_parameters}
\title{Computes Genetic Parameters}
\usage{
gentic_parameters(data, generation, trait)
}
\arguments{
\item{data}{A data frame that contains the generation column and trait column.}

\item{generation}{A vector that contains the generations.}

\item{trait}{A numeric vector that contains the trait value.}
}
\value{
A data frame that summarizes heritabilty, heterosis, pcv, gcv, etc.
}
\description{
Computes Genetic Parameters
}
\examples{
get_summary(data, generation, trait)
}
