% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_scale_joint.R
\name{joint_scale_test}
\alias{joint_scale_test}
\title{Joint Scaling Test}
\usage{
joint_scale_test(data, generation, trait)
}
\arguments{
\item{data}{A data frame that contains the generation column and trait column.}

\item{generation}{A vector that contains the generations.}

\item{trait}{A numeric vector that contains the trait value.}
}
\value{
A list model, genetic parameter and chi square value
}
\description{
Joint Scaling Test
}
\examples{
joint_scale_test(data, generation, trait)
}
