% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_fit.R
\name{perfectf_fit}
\alias{perfectf_fit}
\title{Perfect Fit Test}
\usage{
perfectf_fit(data, generation, trait)
}
\arguments{
\item{data}{A data frame that contains the generation column and trait column.}

\item{generation}{A vector that contains the generations.}

\item{trait}{A numeric vector that contains the trait value.}
}
\value{
A list model, genetic parameter and chi square value
}
\description{
Perfect Fit Test
}
\examples{
perfect_fit(data, generation, trait)
}
