% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_test.R
\name{scale_test}
\alias{scale_test}
\title{Scaling Test (A, B, C, or D)}
\usage{
scale_test(data, generation, trait)
}
\arguments{
\item{data}{A data frame that contains the generation column and trait column.}

\item{generation}{A vector that contains the generations.}

\item{trait}{A numeric vector that contains the trait value.}
}
\value{
A data frame that summarizes the number of observations, mean, mean
of variance, model.
}
\description{
Scaling Test (A, B, C, or D)
}
\examples{
scale_test(data, test="A", generation, trait)
scale_test(data, test="B", generation, trait)
scale_test(data, test="C", generation, trait)
scale_test(data, test="D", generation, trait)
}
